Close DB connections when device activity closes
There are a couple of events that cause the activity to close, eg. changing
screen orientation, closing app, etc. . In such a case we need to close the
db connection to prevent leakage.

=====

[200~W/SQLiteConnectionPool: A SQLiteConnection object for database
'+data+user+0+org_opensuse_dice_yourstorydice+databases+YouStoryDice_db'
was leaked!  Please fix your application to end transactions in progress
properly and to close the database when it is no longer needed.